Changing Variable to Constant
A bug was created when a few calls to the database used a specific
database name instead of the constant in the constants include file.
But when investigating it, it seems that the original developer put the
constant in a variable, which is unnecessary as php constants are
global.  I have replaced all of them for test on dev as this is a
better practice and may have been causing other problems.